requirements.txt
numpy
pandas
scikit-learn
torch
xgboost
joblib
pyyaml
pytest
